整个版本拉了一坨大的上去，加油家豪，我相信你

+++ include/string.h
@@ -1,21 +1,22 @@
 #ifndef _STRING_H
 #define _STRING_H
 
+#include <stddef.h>
 #include "types.h"
 
 // 字符串操作函数
-void *memset(void *s, int c, uint32_t n);
-void *memcpy(void *dest, const void *src, uint32_t n);
-int memcmp(const void *s1, const void *s2, uint32_t n);
+void *memset(void *s, int c, size_t n);
+void *memcpy(void *dest, const void *src, size_t n);
+int memcmp(const void *s1, const void *s2, size_t n);
 size_t strlen(const char *s);
 char *strcpy(char *dest, const char *src);
-char *strncpy(char *dest, const char *src, uint32_t n);
+char *strncpy(char *dest, const char *src, size_t n);
 int strcmp(const char *s1, const char *s2);
-int strncmp(const char *s1, const char *s2, uint32_t n);
+int strncmp(const char *s1, const char *s2, size_t n);
 char *strchr(const char *s, int c);
 char *strrchr(const char *s, int c);
 char *strcat(char *dest, const char *src);
-char *strncat(char *dest, const char *src, uint32_t n);
+char *strncat(char *dest, const char *src, size_t n);
 
 // 字符串转换函数
 int atoi(const char *nptr);
